# cetele/docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Cetele API",
    "description" : "Shop ledger API. Errors are RFC 9457 problem documents with a stable `code`.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "https://cetele.app",
    "description" : "Production"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/v1/shops/{shopId}/invitations" : {
      "post" : {
        "tags" : [ "invitation-controller" ],
        "operationId" : "issue",
        "parameters" : [ {
          "name" : "shopId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateInvitationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IssuedInvitation"
                }
              }
            }
          }
        }
      }
    },
    "/v1/invitations/{code}/accept" : {
      "post" : {
        "tags" : [ "invitation-controller" ],
        "operationId" : "accept",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AcceptedInvitation"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateInvitationRequest" : {
        "type" : "object",
        "properties" : {
          "phone" : {
            "type" : "string",
            "pattern" : "^\\+90[5][0-9]{9}$"
          }
        }
      },
      "IssuedInvitation" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "code" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "AcceptedInvitation" : {
        "type" : "object",
        "properties" : {
          "shopId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "OWNER", "STAFF" ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}